[FIX] alc_product_dimensions: Fix init hook

+++ odoo/addons/alc_product_dimensions/hooks.py
@@ -6,22 +6,31 @@ import logging
 _logger = logging.getLogger(__name__)
 
 
+def column_exists(cr, tablename, columnname):
+    """ Return whether the given column exists. """
+    query = """ SELECT 1 FROM information_schema.columns
+                WHERE table_name=%s AND column_name=%s """
+    cr.execute(query, (tablename, columnname))
+    return cr.rowcount
+
+
 def _update_product_product(cr):
     _logger.info("Update product product from product_template with existing values")
-    cr.execute(
-        """
-        UPDATE
-            product_product pp
-            SET length = pt.length,
-                width = pt.width,
-                height = pt.depth,
-                volume = pt.volume
-        FROM
-            product_template pt
-        WHERE pt.id = pp.product_tmpl_id
+    if column_exists(cr, "product_template", "length"):
+        cr.execute(
+            """
+            UPDATE
+                product_product pp
+                SET length = pt.length,
+                    width = pt.width,
+                    height = pt.depth,
+                    volume = pt.volume
+            FROM
+                product_template pt
+            WHERE pt.id = pp.product_tmpl_id
 
-    """
-    )
+        """
+        )
 
 
 def _update_product_uom(cr):
@@ -62,17 +71,18 @@ def _put_default_uom_id_on_products(cr):
 
 def _delete_useless_columns_product_template(cr):
     _logger.info("Drop useless columns on produt template")
-    cr.execute(
-        """
-        ALTER TABLE
-            product_template
-            DROP COLUMN length,
-            DROP COLUMN width,
-            DROP COLUMN depth,
-            DROP COLUMN volume
+    if column_exists(cr, "product_template", "length"):
+        cr.execute(
+            """
+            ALTER TABLE
+                product_template
+                DROP COLUMN length,
+                DROP COLUMN width,
+                DROP COLUMN depth,
+                DROP COLUMN volume
 
-    """
-    )
+        """
+        )
 
 
 def pre_init_hook(cr):
